+++ alicloud/resource_alicloud_ram_account_alias_test.go
@@ -93,13 +93,8 @@ func TestAccAlicloudRamAccountAlias_basic(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config: testAccRamAccountAliasConfig(acctest.RandIntRange(10000, 999999)),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckRamAccountAliasExists(
-						"alicloud_ram_account_alias.alias", &v),
-					resource.TestMatchResourceAttr(
-						"alicloud_ram_account_alias.alias",
-						"account_alias",
-						regexp.MustCompile("^tf-testaccramaccountalias*")),
+				Check: resource.ComposeTestCheckFunc(testAccCheckRamAccountAliasExists("alicloud_ram_account_alias.alias", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_account_alias.alias", "account_alias", regexp.MustCompile("^tf-testaccramaccountalias*")),
 				),
 			},
 		},