+++ alicloud/resource_alicloud_ram_group_test.go
@@ -118,35 +118,6 @@ func testSweepRamGroups(region string) error {
 	return nil
 }
 
-func TestAccAlicloudRamGroup_basic(t *testing.T) {
-	var v ram.Group
-
-	rand := acctest.RandIntRange(1000000, 99999999)
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-
-		// module name
-		IDRefreshName: "alicloud_ram_group.group",
-
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckRamGroupDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccRamGroupConfig(rand),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckRamGroupExists("alicloud_ram_group.group", &v),
-					resource.TestMatchResourceAttr("alicloud_ram_group.group", "name", regexp.MustCompile("^tf-testAccRamGroupConfig-*")),
-					resource.TestCheckResourceAttr("alicloud_ram_group.group", "comments", "group comments"),
-					resource.TestCheckResourceAttr("alicloud_ram_group.group", "force", "true"),
-				),
-			},
-		},
-	})
-
-}
-
 func TestAccAlicloudRamGroup_default(t *testing.T) {
 	var v ram.Group
 
@@ -167,7 +138,7 @@ func TestAccAlicloudRamGroup_default(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamGroupExists("alicloud_ram_group.group", &v),
 					resource.TestCheckResourceAttr("alicloud_ram_group.group", "name", fmt.Sprintf("tf-testAccRamGroupConfig-%d", randInt)),
-					resource.TestCheckResourceAttr("alicloud_ram_group.group", "comments",""),
+					resource.TestCheckResourceAttr("alicloud_ram_group.group", "comments", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_group.group", "force", "false"),
 				),
 			},