+++ alicloud/resource_alicloud_ram_role_attachment_test.go
@@ -32,14 +32,12 @@ func TestAccAlicloudRamRoleAttachment_basic(t *testing.T) {
 			{
 				Config: testAccRamRoleAttachmentConfig(EcsInstanceCommonTestCase, acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckRamRoleExists(
-						"alicloud_ram_role.role", &role),
-					testAccCheckInstanceExists(
-						"alicloud_instance.instance.0", &instanceA),
-					testAccCheckInstanceExists(
-						"alicloud_instance.instance.1", &instanceB),
-					testAccCheckRamRoleAttachmentExists(
-						"alicloud_ram_role_attachment.attach", &instanceB, &instanceA, &role),
+					testAccCheckRamRoleExists("alicloud_ram_role.role", &role),
+					testAccCheckInstanceExists("alicloud_instance.instance.0", &instanceA),
+					testAccCheckInstanceExists("alicloud_instance.instance.1", &instanceB),
+					testAccCheckRamRoleAttachmentExists("alicloud_ram_role_attachment.attach", &instanceB, &instanceA, &role),
+					resource.TestMatchResourceAttr("alicloud_ram_role_attachment.attach", "role_name", regexp.MustCompile("^tf-testAccRamRoleAttachmentConfig-*")),
+					resource.TestCheckResourceAttrSet("alicloud_ram_role_attachment.attach", instance_ids),
 				),
 			},
 		},
@@ -132,29 +130,24 @@ func testAccRamRoleAttachmentConfig(common string, rand int) string {
 	variable "name" {
 		default = "tf-testAccRamRoleAttachmentConfig-%d"
 	}
-
 	resource "alicloud_instance" "instance" {
 		vswitch_id = "${alicloud_vswitch.default.id}"
 		image_id = "${data.alicloud_images.default.images.0.id}"
-
 		# series III
 		instance_type = "${data.alicloud_instance_types.default.instance_types.0.id}"
 		instance_name = "${var.name}"
 		system_disk_category = "cloud_efficiency"
 		count = 2
-
 		internet_charge_type = "PayByTraffic"
 		internet_max_bandwidth_out = 5
 		security_groups = ["${alicloud_security_group.default.id}"]
 	}
-
 	resource "alicloud_ram_role" "role" {
 	  name = "${var.name}"
 	  services = ["ecs.aliyuncs.com"]
 	  description = "this is a test"
 	  force = true
 	}
-
 	resource "alicloud_ram_role_attachment" "attach" {
 	  role_name = "${alicloud_ram_role.role.name}"
 	  instance_ids = ["${alicloud_instance.instance.*.id}"]