+++ alicloud/resource_alicloud_ram_user_test.go
@@ -144,10 +144,9 @@ func testSweepRamUsers(region string) error {
 	return nil
 }
 
-func TestAccAlicloudRamUser_basic(t *testing.T) {
+func TestAccAlicloudRamUser_withrename(t *testing.T) {
 	var v ram.User
 
-	randInt := acctest.RandIntRange(1000000, 99999999)
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -160,48 +159,22 @@ func TestAccAlicloudRamUser_basic(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig(randInt),
+				Config: testAccRamUserConfig_default(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", "yoyoyo"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
 				),
 			},
 			{
-				Config: testAccRamUserConfig_Update(randInt),
+				Config: testAccRamUserConfig_withrename(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-new-*")),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", "yoyoyo"),
-				),
-			},
-		},
-	})
-
-}
-
-func TestAccAlicloudRamUser_default(t *testing.T) {
-	var v ram.User
-
-	randInt := acctest.RandIntRange(1000000, 99999999)
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-
-		// module name
-		IDRefreshName: "alicloud_ram_user.user",
-
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckRamUserDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccRamUserConfig_default(randInt),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
@@ -214,10 +187,9 @@ func TestAccAlicloudRamUser_default(t *testing.T) {
 
 }
 
-func TestAccAlicloudRamUser_withdispname(t *testing.T) {
+func TestAccAlicloudRamUser_withredispname(t *testing.T) {
 	var v ram.User
 
-	randInt := acctest.RandIntRange(1000000, 99999999)
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -230,11 +202,23 @@ func TestAccAlicloudRamUser_withdispname(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig_withdispname(randInt),
+				Config: testAccRamUserConfig_withdispname(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
+				),
+			},
+			{
+				Config: testAccRamUserConfig_withdispnameupdate(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "new_displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
@@ -249,7 +233,6 @@ func TestAccAlicloudRamUser_withdispname(t *testing.T) {
 func TestAccAlicloudRamUser_withmobile(t *testing.T) {
 	var v ram.User
 
-	randInt := acctest.RandIntRange(1000000, 99999999)
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -262,12 +245,24 @@ func TestAccAlicloudRamUser_withmobile(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig_withmobile(randInt),
+				Config: testAccRamUserConfig_withmobile(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
+				),
+			},
+			{
+				Config: testAccRamUserConfig_withmobileupdate(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-16666666666"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
@@ -281,7 +276,6 @@ func TestAccAlicloudRamUser_withmobile(t *testing.T) {
 func TestAccAlicloudRamUser_withemail(t *testing.T) {
 	var v ram.User
 
-	randInt := acctest.RandIntRange(1000000, 99999999)
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -294,13 +288,25 @@ func TestAccAlicloudRamUser_withemail(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig_withemail(randInt),
+				Config: testAccRamUserConfig_withemail(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
+				),
+			},
+			{
+				Config: testAccRamUserConfig_withemailupdate(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.world@163.com"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", ""),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
 				),
@@ -313,7 +319,6 @@ func TestAccAlicloudRamUser_withemail(t *testing.T) {
 func TestAccAlicloudRamUser_withcomments(t *testing.T) {
 	var v ram.User
 
-	randInt := acctest.RandIntRange(1000000, 99999999)
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -326,7 +331,7 @@ func TestAccAlicloudRamUser_withcomments(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig_withcomments(randInt),
+				Config: testAccRamUserConfig_withcomments(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
 					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
@@ -337,14 +342,25 @@ func TestAccAlicloudRamUser_withcomments(t *testing.T) {
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
 				),
 			},
+			{
+				Config: testAccRamUserConfig_withcommentsupdate(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", ""),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", "RamUser_pls"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "force", "false"),
+				),
+			},
 		},
 	})
 
 }
 
-func TestAccAlicloudRamUser_redisplayname(t *testing.T) {
+func TestAccAlicloudRamUser_multirename(t *testing.T) {
 	var v ram.User
-	randInt := acctest.RandIntRange(1000000, 99999999)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -358,10 +374,10 @@ func TestAccAlicloudRamUser_redisplayname(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig(randInt),
+				Config: testAccRamUserConfig(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -369,10 +385,51 @@ func TestAccAlicloudRamUser_redisplayname(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccRamUserConfig_redisplayname(randInt),
+				Config: testAccRamUserConfig_multirename(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-new-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", "yoyoyo"),
+				),
+			},
+		},
+	})
+
+}
+
+func TestAccAlicloudRamUser_multiredispname(t *testing.T) {
+	var v ram.User
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+
+		// module name
+		IDRefreshName: "alicloud_ram_user.user",
+
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckRamUserDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccRamUserConfig(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
+					resource.TestCheckResourceAttr("alicloud_ram_user.user", "comments", "yoyoyo"),
+				),
+			},
+			{
+				Config: testAccRamUserConfig_multiredispname(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "new_displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -384,9 +441,8 @@ func TestAccAlicloudRamUser_redisplayname(t *testing.T) {
 
 }
 
-func TestAccAlicloudRamUser_remobile(t *testing.T) {
+func TestAccAlicloudRamUser_multiremobile(t *testing.T) {
 	var v ram.User
-	randInt := acctest.RandIntRange(1000000, 99999999)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -400,10 +456,10 @@ func TestAccAlicloudRamUser_remobile(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig(randInt),
+				Config: testAccRamUserConfig(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -411,10 +467,10 @@ func TestAccAlicloudRamUser_remobile(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccRamUserConfig_remobile(randInt),
+				Config: testAccRamUserConfig_multiremobile(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-16666666666"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -426,9 +482,8 @@ func TestAccAlicloudRamUser_remobile(t *testing.T) {
 
 }
 
-func TestAccAlicloudRamUser_newemail(t *testing.T) {
+func TestAccAlicloudRamUser_multireEmail(t *testing.T) {
 	var v ram.User
-	randInt := acctest.RandIntRange(1000000, 99999999)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -442,10 +497,10 @@ func TestAccAlicloudRamUser_newemail(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig(randInt),
+				Config: testAccRamUserConfig(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -453,10 +508,10 @@ func TestAccAlicloudRamUser_newemail(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccRamUserConfig_newemail(randInt),
+				Config: testAccRamUserConfig_multireEmail(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.world@163.com"),
@@ -468,9 +523,8 @@ func TestAccAlicloudRamUser_newemail(t *testing.T) {
 
 }
 
-func TestAccAlicloudRamUser_recomments(t *testing.T) {
+func TestAccAlicloudRamUser_multirecomments(t *testing.T) {
 	var v ram.User
-	randInt := acctest.RandIntRange(1000000, 99999999)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -484,10 +538,10 @@ func TestAccAlicloudRamUser_recomments(t *testing.T) {
 		CheckDestroy: testAccCheckRamUserDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccRamUserConfig(randInt),
+				Config: testAccRamUserConfig(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -495,10 +549,10 @@ func TestAccAlicloudRamUser_recomments(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccRamUserConfig_recomments(randInt),
+				Config: testAccRamUserConfig_multirecomments(acctest.RandIntRange(1000000, 99999999)),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckRamUserExists("alicloud_ram_user.user", &v),
-					resource.TestCheckResourceAttr("alicloud_ram_user.user", "name", fmt.Sprintf("tf-testAccRamUserConfig-%d", randInt)),
+					resource.TestMatchResourceAttr("alicloud_ram_user.user", "name", regexp.MustCompile("^tf-testAccRamUserConfig-*")),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "display_name", "displayname"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "mobile", "86-18888888888"),
 					resource.TestCheckResourceAttr("alicloud_ram_user.user", "email", "hello.uuu@aaa.com"),
@@ -563,32 +617,17 @@ func testAccCheckRamUserDestroy(s *terraform.State) error {
 	return nil
 }
 
-func testAccRamUserConfig(rand int) string {
-	return fmt.Sprintf(`
-	resource "alicloud_ram_user" "user" {
-	  name = "tf-testAccRamUserConfig-%d"
-	  display_name = "displayname"
-	  mobile = "86-18888888888"
-	  email = "hello.uuu@aaa.com"
-	  comments = "yoyoyo"
-	}`, rand)
-}
-
-func testAccRamUserConfig_Update(rand int) string {
-	return fmt.Sprintf(`
-	resource "alicloud_ram_user" "user" {
-	  name = "tf-testAccRamUserConfig-new-%d"
-	  display_name = "displayname"
-	  mobile = "86-18888888888"
-	  email = "hello.uuu@aaa.com"
-	  comments = "yoyoyo"
-	}`, rand)
-}
-
 func testAccRamUserConfig_default(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"
+	}`, rand)
+}
+
+func testAccRamUserConfig_withrename(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-new-%d"
 	}`, rand)
 }
 
@@ -600,11 +639,27 @@ func testAccRamUserConfig_withdispname(rand int) string {
 	}`, rand)
 }
 
+func testAccRamUserConfig_withdispnameupdate(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-%d"
+	  display_name = "new_displayname"
+	}`, rand)
+}
+
 func testAccRamUserConfig_withmobile(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"
 	  mobile = "86-18888888888"
+	}`, rand)
+}
+
+func testAccRamUserConfig_withmobileupdate(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-%d"
+	  mobile = "86-16666666666"
 	}`, rand)
 }
 
@@ -616,6 +671,14 @@ func testAccRamUserConfig_withemail(rand int) string {
 	}`, rand)
 }
 
+func testAccRamUserConfig_withemailupdate(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-%d"
+	  email = "hello.world@163.com"
+	}`, rand)
+}
+
 func testAccRamUserConfig_withcomments(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
@@ -624,7 +687,37 @@ func testAccRamUserConfig_withcomments(rand int) string {
 	}`, rand)
 }
 
-func testAccRamUserConfig_redisplayname(rand int) string {
+func testAccRamUserConfig_withcommentsupdate(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-%d"
+	  comments = "RamUser_pls"
+	}`, rand)
+}
+
+func testAccRamUserConfig(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-%d"
+	  display_name = "displayname"
+	  mobile = "86-18888888888"
+	  email = "hello.uuu@aaa.com"
+	  comments = "yoyoyo"
+	}`, rand)
+}
+
+func testAccRamUserConfig_multirename(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_user" "user" {
+	  name = "tf-testAccRamUserConfig-new-%d"
+	  display_name = "displayname"
+	  mobile = "86-18888888888"
+	  email = "hello.uuu@aaa.com"
+	  comments = "yoyoyo"
+	}`, rand)
+}
+
+func testAccRamUserConfig_multiredispname(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"
@@ -635,7 +728,7 @@ func testAccRamUserConfig_redisplayname(rand int) string {
 	}`, rand)
 }
 
-func testAccRamUserConfig_remobile(rand int) string {
+func testAccRamUserConfig_multiremobile(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"
@@ -646,7 +739,7 @@ func testAccRamUserConfig_remobile(rand int) string {
 	}`, rand)
 }
 
-func testAccRamUserConfig_newemail(rand int) string {
+func testAccRamUserConfig_multireEmail(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"
@@ -657,7 +750,7 @@ func testAccRamUserConfig_newemail(rand int) string {
 	}`, rand)
 }
 
-func testAccRamUserConfig_recomments(rand int) string {
+func testAccRamUserConfig_multirecomments(rand int) string {
 	return fmt.Sprintf(`
 	resource "alicloud_ram_user" "user" {
 	  name = "tf-testAccRamUserConfig-%d"